Add a game dictionary
This dictionary was obtained from
https://www.lexico.com/explore/weird-and-wonderful-words. It will be
used to generate the questions for this game. In the future, the
contents of this file can be replaced with any other dictionary. For
example it can be replaced with a children-specific dictionary or an
architecture-specific jargon dictionary or any other dictionary. The
filename must be maintained. The contents can change.