Defines the play() instance method of GameSession
A call to this method will run the entire game session, asking all the
questions to the user, collecting the answers and updating the data
related to this game session in the instantiated object. For now this
method is only looping through the number of questions and presenting
the same dummy question in each iteration.