Make number of options an instance variable
The GameSession class has now a new instance variable, n_options,
representing the number of options that will be presented to the user
for each question. The idea is to adapt the existing code in a further
commit so that the user will choose the number of options, like
currently the user chooses the number of questions.