Create print_details method
This method was created in the GameSession class. It prints the details
of the game session. The question in the user experience list was also
updated to not contain unnecessary text. It was also added a
`question_number` key to the dictionary that is an element of the user
experience list. All these changes are related to print_details method
so that this method can print the details properly.